Move retry logic to a separate file
