+++ plugins/outputs/cloudwatchlogs/pusher_test.go
@@ -12,14 +12,12 @@ import (
 	"os"
 	"strings"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/influxdata/telegraf/models"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/aws/amazon-cloudwatch-agent/logs"
@@ -797,200 +795,6 @@ func TestResendWouldStopAfterExhaustedRetries(t *testing.T) {
 
 	close(stop)
 	wg.Wait()
-}
-
-func TestChooseRetryWaitStrategy(t *testing.T) {
-	t.Parallel()
-	tests := map[string]struct {
-		err              error
-		expectedStrategy retryWaitStrategy
-	}{
-		"ResourceNotFoundException": {
-			err:              &cloudwatchlogs.ResourceNotFoundException{},
-			expectedStrategy: retryShort,
-		},
-		"InvalidSequenceTokenException": {
-			err:              &cloudwatchlogs.InvalidSequenceTokenException{},
-			expectedStrategy: retryShort,
-		},
-		"ServiceUnavailableException": {
-			err:              &cloudwatchlogs.ServiceUnavailableException{},
-			expectedStrategy: retryLong,
-		},
-		"ThrottlingException": {
-			err:              &cloudwatchlogs.ThrottlingException{},
-			expectedStrategy: retryLong,
-		},
-		"500 - InternalFailure": {
-			err:              awserr.NewRequestFailure(awserr.New("500", "InternalFailure", nil), 500, ""),
-			expectedStrategy: retryLong,
-		},
-		"503 - ServiceUnavailable": {
-			err:              awserr.NewRequestFailure(awserr.New("503", "ServiceUnavailable", nil), 503, ""),
-			expectedStrategy: retryLong,
-		},
-		"Connection Refused": {
-			err:              awserr.New("SomeError", "connection refused", nil),
-			expectedStrategy: retryLong,
-		},
-		"Connection Refused - syscall": {
-			err:              syscall.ECONNREFUSED,
-			expectedStrategy: retryLong,
-		},
-		"Connection Reset By Peer": {
-			err:              awserr.New("SomeError", "connection reset by peer", nil),
-			expectedStrategy: retryLong,
-		},
-		"Connection Reset By Peer - syscall": {
-			err:              syscall.ECONNRESET,
-			expectedStrategy: retryLong,
-		},
-		"Connection Timeout": {
-			err:              syscall.ETIMEDOUT,
-			expectedStrategy: retryLong,
-		},
-		"Request Timeout": {
-			err:              awserr.New("RequestTimeout", "request timed out", nil),
-			expectedStrategy: retryLong,
-		},
-		"Response Timeout": {
-			err:              awserr.New("ResponseTimeout", "response timed out", nil),
-			expectedStrategy: retryLong,
-		},
-		"Deadline Exceeded": {
-			err:              os.ErrDeadlineExceeded,
-			expectedStrategy: retryLong,
-		},
-		"Other Errors": {
-			err:              errors.New("Unknown Error"),
-			expectedStrategy: retryShort,
-		},
-	}
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			chosen := chooseRetryWaitStrategy(tt.err)
-			require.Equal(t, tt.expectedStrategy, chosen)
-		})
-	}
-}
-
-func TestRetryWaitShort(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		n           int
-		minDuration time.Duration
-		maxDuration time.Duration
-	}{
-		{
-			n:           0,
-			minDuration: 100 * time.Millisecond,
-			maxDuration: 200 * time.Millisecond,
-		},
-		{
-			n:           1,
-			minDuration: 200 * time.Millisecond,
-			maxDuration: 400 * time.Millisecond,
-		},
-		{
-			n:           2,
-			minDuration: 400 * time.Millisecond,
-			maxDuration: 800 * time.Millisecond,
-		},
-		{
-			n:           3,
-			minDuration: 800 * time.Millisecond,
-			maxDuration: 1600 * time.Millisecond,
-		},
-		{
-			n:           4,
-			minDuration: 1600 * time.Millisecond,
-			maxDuration: 3200 * time.Millisecond,
-		},
-		{
-			n:           5,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           6,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           7,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%d", tt.n), func(t *testing.T) {
-			for _ = range 1000 {
-				duration := retryWaitShort(tt.n)
-				assert.GreaterOrEqual(t, duration, tt.minDuration, "retryWait(%v) should be greater than or equal to %v", tt.n, tt.minDuration)
-				assert.LessOrEqual(t, duration, tt.maxDuration, "retryWait(%v) should be less than or equal to %v", tt.n, tt.maxDuration)
-			}
-		})
-	}
-}
-
-func TestRetryWaitLong(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		n           int
-		minDuration time.Duration
-		maxDuration time.Duration
-	}{
-		{
-			n:           0,
-			minDuration: 1 * time.Second,
-			maxDuration: 2 * time.Second,
-		},
-		{
-			n:           1,
-			minDuration: 2 * time.Second,
-			maxDuration: 4 * time.Second,
-		},
-		{
-			n:           2,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           3,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           4,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           5,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           6,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-		{
-			n:           7,
-			minDuration: 30 * time.Second,
-			maxDuration: 1 * time.Minute,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%d", tt.n), func(t *testing.T) {
-			for _ = range 1000 {
-				duration := retryWaitLong(tt.n)
-				assert.GreaterOrEqual(t, duration, tt.minDuration, "retryWait(%v) should be greater than or equal to %v", tt.n, tt.minDuration)
-				assert.LessOrEqual(t, duration, tt.maxDuration, "retryWait(%v) should be less than or equal to %v", tt.n, tt.maxDuration)
-			}
-		})
-	}
 }
 
 func testPreparation(retention int, s *svcMock, flushTimeout time.Duration, retryDuration time.Duration) (chan struct{}, *pusher) {
